auth_module/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
redis>=4.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
